+++ apps/web/src/app/accounts/trial-initiation/confirmation-details.component.ts
