+++ apps/web/src/app/shared/loose-components.module.ts
@@ -9,6 +9,7 @@ import { LockComponent } from "../accounts/lock.component";
 import { LoginComponent } from "../accounts/login.component";
 import { RecoverDeleteComponent } from "../accounts/recover-delete.component";
 import { RecoverTwoFactorComponent } from "../accounts/recover-two-factor.component";
+import { RegisterFormModule } from "../accounts/register-form/register-form.module";
 import { RegisterComponent } from "../accounts/register.component";
 import { RemovePasswordComponent } from "../accounts/remove-password.component";
 import { SetPasswordComponent } from "../accounts/set-password.component";
@@ -28,7 +29,6 @@ import { FrontendLayoutComponent } from "../layouts/frontend-layout.component";
 import { NavbarComponent } from "../layouts/navbar.component";
 import { UserLayoutComponent } from "../layouts/user-layout.component";
 import { OrganizationCreateModule } from "../modules/organizations/create/organization-create.module";
-import { RegisterFormModule } from "../modules/register-form/register-form.module";
 import { VaultFilterModule } from "../modules/vault-filter/vault-filter.module";
 import { OrganizationBadgeModule } from "../modules/vault/modules/organization-badge/organization-badge.module";
 import { OrganizationLayoutComponent } from "../organizations/layouts/organization-layout.component";
